+++ packages/ses-integration-test/test/sanity.test.js
@@ -5,12 +5,8 @@ import "ses";
 
 test("sanity", t => {
   t.equal(lockdown(), true, "lockdown runs successfully");
-  const c = new Compartment();
+  const c = new Compartment({ abc: 456 });
   t.equal(c.evaluate("123"), 123, "simple evaluate succeeds");
-  t.equal(
-    c.evaluate("abc", { endowments: { abc: 456 } }),
-    456,
-    "endowment succeeds"
-  );
+  t.equal(c.evaluate("abc"), 456, "endowment succeeds");
   t.end();
 });